% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{checkRootname}
\alias{checkRootname}
\title{Check if files exist}
\usage{
checkRootname(rootname, software = "plink", suffix = "all")
}
\arguments{
\item{rootname}{}

\item{software}{Either "plink" or "gcta"}

\item{suffix}{Either "all" or a specific suffic}
}
\value{
Character string
}
\description{
Either checks of all files for a rootname exist, or returns a specific suffix
}

