% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mr.R
\name{eggers_regression}
\alias{eggers_regression}
\title{Egger's regression for Mendelian randomisation}
\usage{
eggers_regression(b_exp, b_out, se_exp, se_out, bootstrap = NULL)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{bootstrap=NULL}{Number of bootstraps to estimate standard error. If NULL then don't use bootstrap}
}
\value{
List of results from MR Egger
}
\description{
Egger's regression for Mendelian randomisation
}

