% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regression.R
\name{est_se}
\alias{est_se}
\title{Estimate standard error of regression coefficient}
\usage{
est_se(b, n, vx, vy)
}
\arguments{
\item{b}{Effect size}

\item{n}{Sample size}

\item{vx}{Variance of x}

\item{vy}{variance of y}
}
\value{
Expected standard error of b
}
\description{
Estimate standard error of regression coefficient
}

