% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/regression.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{Get the summary stats for use in 2 sample MR}
\usage{
get_summary_stats(y, x)
}
\arguments{
\item{y}{vector of dependant variable}

\item{x}{matrix of independent variables}
}
\value{
data frame of effects and standard errors
}
\description{
Get the summary stats for use in 2 sample MR
}

