% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{makeGRMmatrix}
\alias{makeGRMmatrix}
\title{Convert long GRM format to matrix format}
\usage{
makeGRMmatrix(grm)
}
\arguments{
\item{grm}{Result from readGRM}
}
\value{
Matrix of n x n
}
\description{
Convert long GRM format to matrix format
}

