% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Manhattan Plot}
\usage{
manhattanPlot(p, chr, pos, filename = NULL, width = 15, height = 7,
  threshold = -log10(0.05/1e+06))
}
\arguments{
\item{p}{P values}

\item{chr}{Chromosomes (won't handle X, must not be factors)}

\item{pos}{Physical position}

\item{filename=NULL}{If specified will print to file}

\item{width=15}{}

\item{height=7}{}

\item{threshold=-log10(0.05/1000000)}{For thresholds}
}
\description{
Manhattan Plot
}

