% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{qqplotpval}
\alias{qqplotpval}
\title{QQ plot pvalues}
\usage{
qqplotpval(P, filename = NULL)
}
\arguments{
\item{P}{array of pvalues}

\item{filename=NULL}{If specified the plot will be printed to png}
}
\description{
Creates a qq plot using GenABEL's function. Calculates lambda and prints to title
}

