% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{read_grm}
\alias{read_grm}
\title{Read binary GCTA GRM}
\usage{
read_grm(rootname)
}
\arguments{
\item{rootname}{Name of files. e.g. if the GRM files are data.grm.bin, data.grm.id and data.grm.N.bin then supply "data" as the argument}
}
\value{
List with elements:
        grm: data frame of GRM in long format
        id: data frame of IDs
}
\description{
<full description>
}

