% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{read_plink_linear}
\alias{read_plink_linear}
\title{Read output from plink --linear}
\usage{
read_plink_linear(filename, h = TRUE)
}
\arguments{
\item{filename}{}

\item{h}{Is there a header}
}
\value{
Data.frame
}
\description{
Read output from plink --linear
}

