% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{readGRM}
\alias{readGRM}
\title{Read binary GRM files into R}
\usage{
readGRM(rootname)
}
\arguments{
\item{rootname}{}
}
\value{
List of GRM and id data frames
}
\description{
Read binary GRM files into R
}

