% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{readPlinkLinear}
\alias{readPlinkLinear}
\title{Read output from plink --linear}
\usage{
readPlinkLinear(filename, h = TRUE)
}
\arguments{
\item{filename}{}

\item{h}{Is there a header}
}
\value{
Data.frame
}
\description{
Read output from plink --linear
}

