% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mr.R
\name{two_sample_iv_ml}
\alias{two_sample_iv_ml}
\title{Perform 2 sample IV}
\usage{
two_sample_iv_ml(b_exp, b_out, se_exp, se_out, n = 10000)
}
\arguments{
\item{b_exp}{Vector of genetic effects on exposure}

\item{b_out}{Vector of genetic effects on outcome}

\item{se_exp}{Standard errors of genetic effects on exposure}

\item{se_out}{Standard errors of genetic effects on outcome}

\item{n=10000}{Sample size}
}
\value{
List of results from 2 sample IV
}
\description{
Perform 2 sample IV
}

