% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_write_formats.R
\name{write_grm}
\alias{write_grm}
\title{Write readGRM style output back to binary GRM for use with GCTA}
\usage{
write_grm(grm, rootname)
}
\arguments{
\item{grm}{Output from \link{readGRM}}

\item{rootname}{}
}
\description{
Write readGRM style output back to binary GRM for use with GCTA
}

