% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{writeGRM}
\alias{writeGRM}
\title{Write readGRM style output back to binary GRM for use with GCTA}
\usage{
writeGRM(grm, rootname)
}
\arguments{
\item{grm}{Output from \link{readGRM}}

\item{rootname}{}
}
\description{
Write readGRM style output back to binary GRM for use with GCTA
}

